rewrite to parallel cq-meshing and refinement
This uses python multiprocessing Pool, which in turn meant
we need to have required parameters as class parameters and methods.
This to avoid serialization/pickling of objects which is not allowed.
Limitation is that you can then only have a single instance of the mesher
object, but this is reasonable.